V.0.0.8 (#27)
Realise 0.0.8 of the Profile Management EcoSystem (PMEC)
Sources: https://github.com/Robin8Put/pmes.

Profile Data Management System
The following methods were implemented:
- sell access to the profile is in the process
- edit profile description
- edit profile's selling access price
- edit profile's selling rights price
- add freeze to balance when the offer for buying profile was made
- review system: the only user's who bought profile could leave a review to it

Implementation of multichain with Ethereum is in the progress.

JS client for implementation sell access to the content and generation of access string based on the public key were done. Its integration to the PMES frontend is in the process.